+++ NEWS.md
@@ -1,3 +1,10 @@
+# nodbi 0.15.0
+
+* Add support for MariaDB
+* Add user option `jsonlite.pagesize` for `jsonlite::stream_{in,out}()`
+* Increase default value of `pagesize` for `jsonlite::stream_{in,out}()`
+* Documentation update
+
 # nodbi 0.14.0
 
 * Fix `docdb_query` for `src_sqlite` when using `$in` with strings in query